requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
requests==2.31.0
python-magic-bin==0.4.14
openai==1.6.1
anthropic==0.8.1
firecrawl-py==0.0.5
beautifulsoup4==4.12.2
lxml==4.9.3
